+++ rpc/src/androidMain/kotlin/com/tneff/cyppie/rpc/HttpClientFactory.android.kt
@@ -1,0 +1,8 @@
+package com.tneff.cyppie.rpc
+
+import io.ktor.client.HttpClient
+import io.ktor.client.HttpClientConfig
+import io.ktor.client.engine.okhttp.OkHttp
+
+internal actual fun platformHttpClient(block: HttpClientConfig<*>.() -> Unit): HttpClient =
+    HttpClient(OkHttp) { block() }